Get hibernate working saving tx_data.

+++ src/main/java/txdata/TxDataController.java
@@ -1,6 +1,6 @@
 package txdata;
 
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class TxDataController {
 
     private static final String template = "Hello, %s!";
-    private final AtomicLong counter = new AtomicLong();
+    private final AtomicInteger counter = new AtomicInteger();
 
     @RequestMapping("/api/tx-data")
     public TxData greeting(@RequestParam(value="name", defaultValue="World") String name) {
